+++ src/main/java/com/garbagemule/MobArena/ArenaListener.java
@@ -26,7 +26,9 @@ import org.bukkit.block.Sign;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.AbstractHorse;
 import org.bukkit.entity.AnimalTamer;
+import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Horse;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Ocelot;
@@ -69,6 +71,7 @@ import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.entity.PotionSplashEvent;
 import org.bukkit.event.hanging.HangingBreakEvent;
 import org.bukkit.event.player.PlayerAnimationEvent;
+import org.bukkit.event.player.PlayerArmorStandManipulateEvent;
 import org.bukkit.event.player.PlayerBucketEmptyEvent;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
 import org.bukkit.event.player.PlayerDropItemEvent;
@@ -454,6 +457,10 @@ public class ArenaListener
             return;
         }
 
+        if (arena.inEditMode() && event.getEntityType() == EntityType.ARMOR_STAND) {
+            return;
+        }
+
         if (event.getSpawnReason() != SpawnReason.CUSTOM) {
             if (event.getSpawnReason() == SpawnReason.BUILD_IRONGOLEM || event.getSpawnReason() == SpawnReason.BUILD_SNOWMAN) {
                 monsters.addGolem(event.getEntity());
@@ -717,6 +724,9 @@ public class ArenaListener
         else if (damagee instanceof Ocelot && arena.hasPet(damagee)) {
             onPetDamage(event, (Ocelot) damagee, damager);
         }
+        else if (damagee instanceof ArmorStand) {
+            onArmorStandDamage(event);
+        }
         // Mount
         else if (damagee instanceof AbstractHorse && monsters.hasMount(damagee)) {
             onMountDamage(event, (Horse) damagee, damager);
@@ -770,6 +780,12 @@ public class ArenaListener
 
     private void onPetDamage(EntityDamageEvent event, Ocelot pet, Entity damager) {
         event.setCancelled(true);
+    }
+
+    private void onArmorStandDamage(EntityDamageEvent event) {
+        if (protect && !arena.inEditMode() && region.contains(event.getEntity().getLocation())) {
+            event.setCancelled(true);
+        }
     }
 
     private void onMountDamage(EntityDamageEvent event, Horse mount, Entity damager) {
@@ -1098,6 +1114,12 @@ public class ArenaListener
         else if (event.getClickedBlock().getState() instanceof Sign) {
             Sign sign = (Sign) event.getClickedBlock().getState();
             handleSign(sign, p);
+        }
+    }
+
+    public void onPlayerArmorStandManipulate(PlayerArmorStandManipulateEvent event) {
+        if (protect && !arena.inEditMode() && region.contains(event.getRightClicked().getLocation())) {
+            event.setCancelled(true);
         }
     }